// src/app/interfaces/suppliers.ts
export interface Supplier {
  _id: string;  
  businessName: string;
  name: string;
  lastName: string;
  birthDate: string;
  mobile: string;
  sdi: string;
  status: number;
  email: string;
  vatNumber: string;
  agentName: string;
  address: string;
  zipCode: string;
  province: string;
  city: string;
  pwd: string;
}